src/test/java/dk/dma/embryo/util/DateTimeConverterTest.java
/* Copyright (c) 2011 Danish Maritime Authority
 *
 * This library is free software; you can redistribute it and/or
 * modify it under the terms of the GNU Lesser General Public
 * License as published by the Free Software Foundation; either
 * version 3 of the License, or (at your option) any later version.
 *
 * This library is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
 * Lesser General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with this library.  If not, see <http://www.gnu.org/licenses/>.
 */
package dk.dma.embryo.util;

import static org.junit.Assert.assertEquals;

import org.joda.time.DateTime;
import org.junit.Test;

/**
 * @author Jesper Tejlgaard
 */
public class DateTimeConverterTest {

    @Test
    public void test() {

        DateTime dateTime = new DateTime(1392298200000L);

        String str = DateTimeConverter.getDateTimeConverter().toStringMedium(dateTime);

        assertEquals("13-02-2014 13:30:00", str);
    }

}
